Doc updates and config restructuring
