+++ packages/cache/package.json
@@ -16,7 +16,6 @@
     "watch": "tsc -b -w"
   },
   "dependencies": {
-    "@azure/core-http": "^3.0.0",
     "@azure/storage-blob": "^12.15.0",
     "backfill-config": "^6.6.1",
     "backfill-logger": "^5.3.0",
