+++ packages/config/src/azureBlobCacheConfig.ts
@@ -1,0 +1,91 @@
+import { cacheConfigEnvNames as envNames } from "./cacheConfig";
+
+export type AzureBlobCacheStorageConfig = {
+  provider: "azure-blob";
+  options: AzureBlobCacheStorageOptions;
+};
+
+export type AzureBlobCacheStorageOptions =
+  | AzureBlobCacheStorageConnectionStringOptions
+  | AzureBlobCacheStorageContainerClientOptions;
+
+export type AzureBlobCacheStorageConnectionStringOptions = {
+  /** Retrieve this from the Azure portal for the blob storage. */
+  connectionString: string;
+  /** Name of the blob storage container. */
+  container: string;
+  /** Max size of a single package cache, in bytes. */
+  maxSize?: number;
+  /** One of the credential types from `@azure/identity`. */
+  credential?: TokenCredentialLike;
+};
+
+export type AzureBlobCacheStorageContainerClientOptions = Pick<
+  AzureBlobCacheStorageConnectionStringOptions,
+  "maxSize"
+> & {
+  /**
+   * Pre-created `ContainerClient` from `@azure/storage-blob`, as returned by
+   * `BlobServiceClient#getContainerClient`.
+   */
+  containerClient: ContainerClientLike;
+};
+
+// The ___Like types prevent pulling in large trees of azure type dependencies,
+// particularly in the lage CLI's types rollup (which includes backfill cache
+// options). The objects should be generated by the relevant Azure SDKs, so
+// they'll automatically match the full signatures. If the stub signatures
+// become mismatched, it should show up when type checking the backfill-cache
+// package that uses the Azure SDKs.
+
+/**
+ * `ContainerClient` from `@azure/storage-blob`, as returned by
+ * `BlobServiceClient#getContainerClient`.
+ */
+type ContainerClientLike = {
+  getBlobClient: (hash: string) => unknown;
+};
+
+/** One of the credential types from `@azure/identity`. */
+type TokenCredentialLike = {
+  getToken: (scopes: string | string[], options?: unknown) => Promise<any>;
+};
+
+export function getAzureBlobConfigFromSerializedOptions(
+  options: string
+): AzureBlobCacheStorageConfig {
+  let parsedOptions: AzureBlobCacheStorageConnectionStringOptions;
+  try {
+    parsedOptions = JSON.parse(options);
+  } catch {
+    throw new Error(
+      `Could not parse ${envNames.cacheProviderOptions} as JSON:\n"${options}"`
+    );
+  }
+
+  if (
+    typeof parsedOptions.connectionString !== "string" ||
+    typeof parsedOptions.container !== "string"
+  ) {
+    throw new Error(
+      `Invalid ${envNames.cacheProviderOptions} for ${envNames.cacheProvider}="azure-blob":\n` +
+        `Expected: object with string values for keys "connectionString", "container"\n` +
+        `Received: "${options}"`
+    );
+  }
+
+  if (
+    parsedOptions.maxSize !== undefined &&
+    typeof parsedOptions.maxSize !== "number"
+  ) {
+    throw new Error(
+      `Invalid ${envNames.cacheProviderOptions} for ${envNames.cacheProvider}="azure-blob":` +
+        `expected "maxSize" property to be unspecified or a number; received: ${parsedOptions.maxSize}`
+    );
+  }
+
+  return {
+    provider: "azure-blob",
+    options: parsedOptions,
+  };
+}
